% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_subgraphs.R
\name{color_subgraphs}
\alias{color_subgraphs}
\title{Color subgraphs}
\usage{
color_subgraphs(
  DT,
  new_col = "color",
  verbose = getOption("hutilsc.verbose", FALSE)
)
}
\arguments{
\item{DT}{A \code{data.table}, whose keys represent the edges of a graph.
As the graph is undirected, each edge must connect a node to a higher node.}

\item{new_col}{\code{character(1)} The new column name for the coloring.}

\item{verbose}{Whether to display verbose info.}
}
\value{
\code{DT} with an extra integer column, identifying each subgraph.
}
\description{
Color subgraphs
}
\examples{
library(data.table)
DT <- data.table(x = rpois(30, 15), y = rpois(30, 15))
DT <- DT[x != y]
DT[, x1 := pmin(x, y)][, x2 := pmax(x, y)]
DT[, x := NULL][, y := NULL]
setkey(DT, x1, x2)
color_subgraphs(DT)

#                       #   #  ##  ##  ##  ## 
DT <- data.table(x = c(1L, 1L, 2L, 4L, 8L, 9L),
                 y = c(2L, 3L, 4L, 5L, 9L, 10L))
setkey(DT, x, y)
color_subgraphs(DT)


}
