% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_via_edges.R
\name{distance_via_edges}
\alias{distance_via_edges}
\title{Distance between nodes given edges}
\usage{
distance_via_edges(x, y, dt_edges)
}
\arguments{
\item{x, y}{Two vectors of nodes.}

\item{dt_edges}{The edges of the graph, given as a keyed data table.
Two nodes \code{n1, n2} are connected if there exists a row in \code{dt_edges}
with both \code{n1} and \code{n2} in the first two columns.}
}
\value{
A vector the same length as \code{x}, giving the integer
distance between each element of \code{x} and \code{y}.
}
\description{
Distance between nodes given edges
}
\examples{

library(data.table)
edges <- data.table(orig = 1:5, dest = 2:6)
setkey(edges, orig, dest)
distance_via_edges(1L, 2L, edges)
distance_via_edges(1L, 3L, edges)
distance_via_edges(1L, 5L, edges)


}
