% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encode3202.R
\name{Encode3202}
\alias{Encode3202}
\alias{Decode3202}
\alias{Validate3202}
\alias{Lookup4}
\alias{names2int}
\title{Encode}
\usage{
Encode3202(x)

Decode3202(enc)

Validate3202(x)

Lookup4(enc)

names2int(x, y)
}
\arguments{
\item{x, y}{\code{character(n)} Vectors to encode.}

\item{enc}{\code{integer(n)} Encoded vectors to decode.}
}
\value{
\code{Encode3202} returns an integer based on the digits following
the prefix \code{"3202"}. \code{Decode3202} is the reverse process.

\code{names2int} takes two character vectors of equal length and returns
an integer based on the first two digits. \code{Lookup4} is the inverse
process, returning the first two characters of each original vector
in \code{AaAa} case.
}
\description{
Encode
}
