% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_fwc.R
\name{ensure_fwc}
\alias{ensure_fwc}
\title{Ensure character vector is fixed width}
\usage{
ensure_fwc(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
If each element of \code{x} has the same number of characters,
\code{x} is returned.

Otherwise, each element narrower of the maximal width is padded with
leading zeroes.
}
\description{
Ensure character vector is fixed width
}
