% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAbsent.R
\name{findAbsent}
\alias{findAbsent}
\title{Find singleton missing value}
\usage{
findAbsent(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{An integer vector, a proper subset of \code{1:N} with
\code{N == length(x) + 1}.}
}
\description{
Demo of \code{xor} detection of absent value.
}
\examples{
A <- sample.int(1003, size = 1002)
setdiff(1:1003, A)
findAbsent(A)

}
