% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine_dist_c.R
\name{haversine_dist_c}
\alias{haversine_dist_c}
\title{Haversine distance}
\usage{
haversine_dist_c(lat1, lon1, lat2, lon2, use_float = TRUE)
}
\arguments{
\item{lat1, lon1, lat2, lon2}{\describe{
\item{\code{double(N)}}{Coordinates between which the distances are required.}
}}

\item{use_float}{Whether to allow a more imprecise square root function.
Set to \code{TRUE} for performance; set to \code{FALSE} for accuracy. (Though
note well that even setting \code{TRUE} will not be totally accurate
in measuring the real-world distance between two coordinates.)}
}
\description{
Haversine distance
}
