% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutilsc-package.R
\docType{package}
\name{hutilsc-package}
\alias{hutilsc}
\alias{hutilsc-package}
\title{Package of miscellaneous functions in C}
\description{
Sister of the 'hutilscpp' package, designed for better install time.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
