% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar.R
\name{hutilsc_nchar}
\alias{hutilsc_nchar}
\alias{range_nchar}
\title{Number of characters}
\usage{
hutilsc_nchar(x)

range_nchar(x)
}
\arguments{
\item{x}{A character vector. Integer vectors are also supported for convenience.}
}
\value{
\describe{
\item{\code{hutilsc_nchar}}{A faster version of base::nchar(x) but only
measures bytes.}
\item{\code{range_nchar}}{An length-2 integer vector, equivalent to
\code{range(nchar(x, type = "bytes", keepNA = FALSE))}.}
}
}
\description{
Various enhancements of \code{nchar} and related operations.
}
