% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary_exp.R
\name{is_binary_expr}
\alias{is_binary_expr}
\title{Is a call a binary expression}
\usage{
is_binary_expr(sexp)
}
\arguments{
\item{sexp}{A call, likely the result of \code{substitute(expr)}
within a function on a user-supplied \code{expr}.}
}
\value{
\code{TRUE} if a binary expression.
}
\description{
Is a call a binary expression
}
