% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_constant.R
\name{is_constant}
\alias{is_constant}
\title{Is a vector constant}
\usage{
is_constant(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{An atomic vector.}

\item{nThread}{Number of threads to use.}
}
\description{
Is a vector constant
}
