% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_from_edges.R
\name{is_valid_path}
\alias{is_valid_path}
\alias{which_reached_between}
\alias{common_contacts}
\title{Path Between Nodes From List of Edges}
\usage{
is_valid_path(path, Edges)

which_reached_between(a, b, Edges)

common_contacts(a, b, Edges, len = 3L)
}
\arguments{
\item{path}{Integer vector, a candidate path.}

\item{Edges}{A list of two integer vectors, giving each edge of the graph.}

\item{a, b}{Scalar integers. The origin and destination of a path.}

\item{len}{For \code{common_contacts}, the length of the path to consider.}
}
\value{
If a path from \code{orig} to \code{dest} exists, \code{is_valid_path}
returns \code{TRUE}.

\code{common_contacts} returns the nodes that are between the nodes.
}
\description{
Path Between Nodes From List of Edges
}
