% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuniqueN.R
\name{kuniqueN}
\alias{kuniqueN}
\title{The number of unique elements in a key}
\usage{
kuniqueN(x, k = 1L, silent = FALSE)
}
\arguments{
\item{x}{Either a \code{data.table} or a sorted vector.}

\item{k}{An integer or string denoting for which key the number of unique
elements is sought.
By default, the first key of \code{x} is chosen. If \code{x} is an atomic
vector, \code{k} is ignored (as \code{x} is assumed to be the elements of the
key itself).}
}
\description{
The number of unique elements in a key
}
