% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Minimum and maximum}
\usage{
minmax(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{A vector for which the minimum and maximum are desired.}

\item{nThread}{An integer, the number of threads to use.}
}
\description{
Minimum and maximum
}
