% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_from_edges.R
\name{n_paths}
\alias{n_paths}
\alias{n_paths_stv}
\title{Number of paths between vertices}
\usage{
n_paths_stv(s = NULL, v = NULL, t = NULL, Edges, nPaths = NULL)
}
\arguments{
\item{s}{\code{int} A vertex in Edges. The origin of the paths.}

\item{v}{\code{int} Optionally a vertex between \code{s} and \code{t}.}

\item{t}{\code{int} A vertex in Edges. The destination.}

\item{Edges}{A \code{data.table} designating the edgelist.}
}
\description{
Number of paths between vertices
}
