% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_paths_between_given_dist.R
\name{n_paths_between_given_dist}
\alias{n_paths_between_given_dist}
\title{Number of paths when distance is known (or knowable)}
\usage{
n_paths_between_given_dist(Edges, MoltenDistances = NULL, return_dt = TRUE)
}
\arguments{
\item{Edges}{A \code{data.table} giving the edges.}

\item{MoltenDistances}{A \code{data.table} indicating the distance between
each node in \code{Edges}.}

\item{return_dt}{\code{TRUE | FALSE} Should a \code{data.table} be returned?
If \code{TRUE}, it will be appended to \code{MoltenDistances}.
If \code{FALSE}, a vector indicating the number of (shortest) paths between
each node.}
}
\description{
Number of paths when distance is known (or knowable)
}
