% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_paths_svt.R
\name{n_paths_svt}
\alias{n_paths_svt}
\title{Number of paths between vertices}
\usage{
n_paths_svt(s = NULL, v = NULL, t = NULL, Edges, nPaths = NULL)
}
\arguments{
\item{s}{\code{int} A vertex in Edges. The origin of the paths.}

\item{v}{\code{int} Optionally a vertex between \code{s} and \code{t}. If \code{NULL},
\code{n_paths_svt} returns the number of shortest paths between \code{s} and \code{t}.}

\item{t}{\code{int} A vertex in Edges. The destination.}

\item{Edges}{A \code{data.table} designating the edgelist.}

\item{nPaths}{A \code{data.table} indicating the number of paths between
\emph{pairs} of each vertex in \code{Edges}.}
}
\description{
Number of paths between vertices
}
