% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_paths_svt0.R
\name{n_paths_svt0}
\alias{n_paths_svt0}
\title{Number of paths between vertices}
\usage{
n_paths_svt0(
  s,
  v = NULL,
  t,
  Edges,
  weight_col = 3L,
  MoltenDistances = NULL,
  double_edges = TRUE,
  nThread = getOption("hutilsc.nThread", 1L)
)
}
\arguments{
\item{s}{\code{int} A vertex in Edges. The origin of the paths.}

\item{v}{\code{int} Optionally a vertex between \code{s} and \code{t}. If \code{NULL},
\code{n_paths_svt} returns the number of shortest paths between \code{s} and \code{t}.}

\item{t}{\code{int} A vertex in Edges. The destination.}

\item{Edges}{A \code{data.table} designating the edgelist.}

\item{weight_col}{Which column of \code{Edges} should be considered the edge
weight? Either an \code{int} or a string.}

\item{MoltenDistances}{A \code{data.table} indicating the distance between
each pair of vertices in \code{Edges}.}

\item{double_edges}{\code{TRUE | FALSE} Whether or not to double the
\code{Edges} before analysing.}

\item{nThread}{Number of threads to use.}
}
\description{
Number of paths between vertices
}
