% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad0.R
\name{pad0}
\alias{pad0}
\title{Pad a character vector with leading zeroes}
\usage{
pad0(x, w = NULL)
}
\arguments{
\item{x}{A character vector.}

\item{w}{The necessary width; strings in \code{x} where \code{nchar(x) < w}
will be padded with the necessary leading zeroes.}
}
\description{
Useful for ensuring an equilength vector.
}
