% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_from_edges.R
\name{path_from_edges}
\alias{path_from_edges}
\title{Path Between Nodes From List of Edges}
\usage{
path_from_edges(
  orig,
  dest,
  Edges,
  color_colname = NULL,
  return. = c("list", "union")
)
}
\arguments{
\item{orig, dest}{Scalar integers, the start and end nodes of the putative path.}

\item{Edges}{A list of two integer vectors, giving each edge of the graph.}
}
\value{
If a path from \code{orig} to \code{dest} exists, the nodes
between
}
\description{
Path Between Nodes From List of Edges
}
