% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_hash.R
\name{pcg_hash}
\alias{pcg_hash}
\title{PCG Hash}
\usage{
pcg_hash(
  n,
  r = NULL,
  raw_result = FALSE,
  nThread = getOption("hutilsc.nThread", 1L)
)
}
\arguments{
\item{n}{Length of the output.}

\item{r}{Integer vector, typically the random seed. Provide an integer vector
as long as \code{.Random.seed} to reproduce results.}

\item{raw_result}{\code{TRUE | FALSE} Whether or not the output should be
type \code{raw}.}

\item{nThread}{Number of threads to use.}
}
\description{
PCG Hash
}
\examples{
pcg_hash(4)


}
