% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgraph.R
\name{qgraph.layout.fruchtermanreingold}
\alias{qgraph.layout.fruchtermanreingold}
\title{Layout of nodes from edge weights}
\usage{
qgraph.layout.fruchtermanreingold(
  edgelist,
  Ef = .subset2(edgelist, key(edgelist)[1]),
  Et = .subset2(edgelist, key(edgelist)[2]),
  weights = rep(1, nrow(edgelist)),
  vcount = length(union(Ef, Et)),
  niter = 500L,
  max.delta = NULL,
  area = NULL,
  cool.exp = 1.5,
  repulse.rad = NULL,
  init = NULL,
  groups = NULL,
  rotation = NULL,
  layout.control = 0.5,
  constraints = NULL,
  round = TRUE,
  digits = 5
)
}
\arguments{
\item{edgelist}{A \code{data.table} giving the edges.}

\item{Ef, Et}{Integer vectors indicating the edges from and to.}

\item{weights, vcount, niter, max.delta, area, cool.exp, repulse.rad, init, groups, rotation, layout.control, constraints, round, digits}{As in the \code{qgraph} equivalent.}
}
\description{
Layout of nodes from edge weights
}
