% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_racf.R
\name{simulate_racf}
\alias{simulate_racf}
\alias{set_epipars}
\title{Simulate transmission through bipartite graph}
\usage{
simulate_racf(
  STP,
  keyz = c("id", "racf_abm"),
  Epi = set_epipars(),
  PatientZeroGroup = NULL,
  PatientZero = sample(nrow(STP), size = 1),
  n_days = 28L,
  nThread = getOption("hutilsc.nThread", 1L)
)

set_epipars(
  incubation_distribution = c("dirac", "pois", "lnorm", "cauchy"),
  incubation_mean = 8,
  incubation_sigma = 0.44,
  v_workplaces = 0.1,
  q_workplace_rate = 1/14,
  r_workplace_rate = 2
)
}
\arguments{
\item{STP}{A data.table with two keys. The first key is the
identifier of the person; the second key is the identifer of the group
of the person.

If STP has a column \code{"isVaccinated"}, it must be a logical column.
If present, it may be used to indicate which rows are vaccinated. If
a row is \code{NA}, it will be replaced with \code{TRUE} or \code{FALSE}
based on the epidemiological parameter \code{Epi$v_workplaces}.}

\item{Epi}{Epidemiological parameters.}
}
\description{
Simulate transmission through bipartite graph
}
