% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortedness.R
\name{sorted}
\alias{sorted}
\alias{isntSorted}
\alias{is_sorted}
\title{Various sorting functions}
\usage{
isntSorted(x)

is_sorted(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{Atomic vector to sort.}

\item{nThread}{\code{integer(1)} Number of threads to use to verify
sortedness.}
}
\description{
Various sorting functions
}
