% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_isna.R
\name{sum_isna}
\alias{sum_isna}
\title{Sum of NA values}
\usage{
sum_isna(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{Integer vector.}

\item{nThread}{A scalar integer, number of threads.}
}
\description{
Sum of NA values
}
