% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_raw.R
\name{sum_raw}
\alias{sum_raw}
\alias{tabulate_raw}
\title{Sum a raw vector}
\usage{
sum_raw(x, nThread = getOption("hutilsc.nThread", 1L))

tabulate_raw(x, nThread = getOption("hutilsc.nThread", 1L))
}
\arguments{
\item{x}{A raw atomic vector.}

\item{nThread}{Number of threads to use.

See also \code{bit::bit}.}
}
\value{
Summarizes a raw vector:
\describe{
\item{\code{sum_raw}}{The sum of \code{x} as if the elements were unsigned integers.
Returns \code{NULL} if \code{x} is not raw.}
\item{\code{tabulate_raw}}{A length-256 integer vector tabulating the number of
raw elements. }
}
}
\description{
In particular when using the \code{unsigned char}
to represent a bool.
}
\examples{
x <- c(as.raw(0), as.raw(1))
sum_raw(x)
tabulate_raw(x)

x <- c(as.raw(0), as.raw(1))
sum_raw(x)
tabulate_raw(x)

}
