% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-edges.R
\name{test_reaches_dest}
\alias{test_reaches_dest}
\title{Edge summaries}
\usage{
test_reaches_dest(a, b, Edges)
}
\arguments{
\item{a, b}{Nodes}

\item{Edges}{A data.table, keyed, of undirected edges.}
}
\description{
Edge summaries
}
