% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_net.R
\name{ego_net}
\alias{ego_net}
\title{Ego graphs from networks}
\usage{
ego_net(v, order = 1L, Edges, u = NULL)
}
\arguments{
\item{v}{Integer vector of vertices}

\item{order}{Single integer, the maximum discrete distance between any vertex
in \code{v}}

\item{Edges}{A \code{data.table} describing unweighted, undirected edges
of a graph.}

\item{u}{Optionally, the integer vector of nodes.}
}
\value{
A \code{data.table} of two columns, the first is a column
of nodes, matching \code{u}, and the second is the order of the
given node relative to \code{v}, or \code{NA} if the order is larger
than \code{v}.
}
\description{
Also known as the neighbourhood, though peculiar to
social networks
}
