% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_from_edges.R
\name{n_paths}
\alias{n_paths}
\alias{n_paths_svt}
\alias{n_paths_between_given_dist}
\alias{n_paths_svt0}
\title{Number of paths between vertices}
\usage{
n_paths_svt(s = NULL, v = NULL, t = NULL, Edges, nPaths = NULL)

n_paths_between_given_dist(Edges, MoltenDistances = NULL, return_dt = TRUE)

n_paths_svt0(
  s,
  v = NULL,
  t,
  Edges,
  weight_col = 3L,
  MoltenDistances = NULL,
  double_edges = TRUE,
  nThread = getOption("hutilsc.nThread", 1L)
)
}
\arguments{
\item{s}{\code{int} A vertex in Edges. The origin of the paths.}

\item{v}{\code{int} Optionally a vertex between \code{s} and \code{t}. If \code{NULL},
\code{n_paths_svt} returns the number of shortest paths between \code{s} and \code{t}.}

\item{t}{\code{int} A vertex in Edges. The destination.}

\item{Edges}{A \code{data.table} designating the edgelist.}
}
\description{
Number of paths between vertices
}
